set individual mount tags for 9p volumes
To be able to mount multiple 9p volumes (especially on VMs) we need to
set individual mount tags for each volume, instead of reusing the same
tag "shared_dir".

This commit sets mount tags for 9p volumes to the corresponding volume
ID, with the exception of the root volume for container edge apps - it
is changed from "shared_dir" to "container_rootfs" for better
readability. The tag of container's rootfs volume needs to be constant
since it's mounted automatically by the initrd script.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>