feat: Add snapshot creation from existing rootfs in domainmgr
This commit adds the functionality to create a snapshot from an existing rootfs in the domainmgr package. It creates a snapshot using the containerd client and the specified snapshotter. The snapshot is created in the "eve-user-apps" namespace and is based on the "xen-tools" snapshot. This feature allows for the creation of snapshots that can be used as a base for user services.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>

feat: Add error handling for existing temporary snapshot in CreateSnapshotFromExistingRootfs

This commit adds error handling for the case when a temporary snapshot already exists in the CreateSnapshotFromExistingRootfs function of the containerd package. If a temporary snapshot with the same name exists, it is removed before creating a new one.It also adds the "gc.root" label to xen-tools snapshot to prevent it from being garbage collected.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>